requirements.txt
aiogram>=3.24.0
aiohttp>=3.13.3
aiofiles>=25.1.0
magic-filter>=1.0.12
pytube==15.0.0
yt-dlp==2024.11.11
python-dotenv==1.0.0
requests==2.31.0
imageio-ffmpeg